Move custom response classes used for POST request to a new module in util to be used in GET response handling by model views
